Refactor server build to TypeScript
