+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "screen-capture",
   "version": "1.0.0",
-  "description": "Сервис записи экрана, трансляции и видеозвонков",
-  "main": "server.js",
+  "description": "\u0421\u0435\u0440\u0432\u0438\u0441 \u0437\u0430\u043f\u0438\u0441\u0438 \u044d\u043a\u0440\u0430\u043d\u0430, \u0442\u0440\u0430\u043d\u0441\u043b\u044f\u0446\u0438\u0438 \u0438 \u0432\u0438\u0434\u0435\u043e\u0437\u0432\u043e\u043d\u043a\u043e\u0432",
+  "main": "server/dist/server.js",
   "scripts": {
-    "prestart": "npm run build",
-    "start": "node server.js",
-    "dev": "cross-env NODE_ENV=development node server.js",
-    "prod": "cross-env NODE_ENV=production node server.js",
+    "prestart": "npm run build && npm run build:server",
+    "start": "node server/dist/server.js",
+    "dev": "npm run build:server && cross-env NODE_ENV=development node server/dist/server.js",
+    "prod": "npm run build:server && cross-env NODE_ENV=production node server/dist/server.js",
     "lint": "npm run lint:server && npm run lint:client",
     "lint:server": "eslint .",
     "lint:client": "npm --prefix frontend install && npm --prefix frontend run lint",
@@ -19,7 +19,8 @@
     "dev:client": "npm --prefix frontend run dev -- --host",
     "test:server": "vitest --run --config vitest.server.config.ts",
     "test:client": "npm --prefix frontend run test",
-    "test": "npm run test:server && npm run test:client"
+    "test": "npm run test:server && npm run test:client",
+    "build:server": "tsc -p server/tsconfig.json"
   },
   "keywords": [],
   "author": "",
@@ -39,6 +40,10 @@
     "globals": "^15.15.0",
     "prettier": "^3.4.2",
     "supertest": "^7.1.0",
-    "vitest": "^2.1.8"
+    "vitest": "^2.1.8",
+    "typescript": "^5.6.3",
+    "@types/express": "^4.17.21",
+    "@types/node": "^20.17.6",
+    "@types/ws": "^8.5.12"
   }
 }
