+++ server/api/calls.test.ts
@@ -1,9 +1,7 @@
 import request from 'supertest';
 import { describe, expect, it } from 'vitest';
 
-import serverModule from '../../server.js';
-
-const { app } = serverModule as { app: import('express').Express };
+import { app } from '#server';
 
 const uuidV4Regex = /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
